Fix Python HTTP app demo
- Rename main file to "app.py" (to match default in Dockerfile)
- Fix code to work under Python 3
