+++ app.py
@@ -1,0 +1,24 @@
+#!/usr/bin/python
+
+from http.server import BaseHTTPRequestHandler, HTTPServer
+
+PORT_NUMBER = 8080
+
+class MyHandler(BaseHTTPRequestHandler):
+
+  def do_GET(self):
+    """Handler for GET requests"""
+    self.send_response(200)
+    self.send_header('Content-type','image/png')
+    self.end_headers()
+    with open('logo.png', 'rb') as f:
+      self.wfile.write(f.read())
+
+try:
+  server = HTTPServer(('', PORT_NUMBER), MyHandler)
+  print('Started httpserver on port', PORT_NUMBER)
+  server.serve_forever()
+
+except KeyboardInterrupt:
+  server.server_close()
+  print('Stopping server')
